+++ src/lang/builtins_thread.ml
@@ -73,6 +73,39 @@ let () =
       Lang.unit)
 
 let () =
+  let fun_t =
+    Lang.fun_t
+      [
+        (false, "backtrace", Lang.string_t);
+        (false, "queue_name", Lang.string_t);
+        (false, "", Builtins_error.Error.t);
+      ]
+      Lang.unit_t
+  in
+  add_builtin "thread.on_error" ~cat:Liq
+    ~descr:
+      "Register a function to be called when an error of the given kind is \
+       raised in a asynchronous thread."
+    [("", Builtins_error.Error.t, None, None); ("", fun_t, None, None)]
+    Lang.unit_t (fun p ->
+      let err = Builtins_error.Error.of_value (Lang.assoc "" 1 p) in
+      let fn = Lang.assoc "" 2 p in
+      let handler ~bt ~name = function
+        | Lang_values.(Runtime_error { kind; msg; _ })
+          when kind = err.Builtins_error.kind ->
+            let error = Builtins_error.(Error.to_value { kind; msg }) in
+            let bt = Lang.string bt in
+            let name = Lang.string name in
+            ignore
+              (Lang.apply fn
+                 [("backtrace", bt); ("queue_name", name); ("", error)]);
+            true
+        | _ -> false
+      in
+      Queue.push handler Tutils.error_handlers;
+      Lang.unit)
+
+let () =
   let t = Lang.univ_t () in
   add_builtin "thread.mutexify" ~cat:Liq
     ~descr: